[TRL-430] refactor: UserPayload 계층 변경
- infrastructure -> application
