+++ src/main/java/com/cosain/trilo/trip/presentation/schedule/ScheduleDeleteController.java
@@ -1,6 +1,6 @@
 package com.cosain.trilo.trip.presentation.schedule;
 
-import com.cosain.trilo.auth.infra.token.UserPayload;
+import com.cosain.trilo.auth.application.token.UserPayload;
 import com.cosain.trilo.auth.presentation.Login;
 import com.cosain.trilo.auth.presentation.LoginUser;
 import com.cosain.trilo.trip.application.schedule.service.ScheduleDeleteService;
