+++ src/main/java/com/cosain/trilo/user/presentation/UserRestController.java
@@ -1,6 +1,6 @@
 package com.cosain.trilo.user.presentation;
 
-import com.cosain.trilo.auth.infra.token.UserPayload;
+import com.cosain.trilo.auth.application.token.UserPayload;
 import com.cosain.trilo.auth.presentation.Login;
 import com.cosain.trilo.auth.presentation.LoginUser;
 import com.cosain.trilo.user.application.UserService;
